Added Card, Deck, HomeController, Hand, and PlayingField classes.
Implemented methods to alter the text of Hand cards at run time with their given random value. Added variables to track the Rectangles, Text, StackPanes of the hand cards. Attempted to use arrays or lists but was baffled at how difficult it would be to do so. Manual tracking of variables is likely needed in order to move ahead with this project in good time.