Added functionality for the info text.
I realized that it sometimes isn't very obvious what is happening, or if someone just won the whole game or a round, even with the indicators.

I decided to include functionality through SetInfoText(String text) to allow us to add info text. I've also gone ahead and attempted to implement this system whenever an action happens in the game that is important.

I had to change compareScores() in GameLogic to return an integer based on who won, 0 if a draw, 1 if player 1, 2 if player 2, that way I could add the correct Player number to the info text in certain situations.